fix: restore post-merge validation
Entire-Checkpoint: c17b9e00c732
